move usage function to helper
